updated tests/template tags
