+++ django_faker_tests/tests/test_example.py
@@ -1,6 +1,54 @@
+from faker import Faker
 from django import test
 from django_faker import DjangoFaker
-from ..models import Assessment, Question, Answer, AnswerSubmission, Company
+from django_faker.populator import Populator
+from ..models import Assessment, Question, Answer, AnswerSubmission, Company, Game, Player, Action
+
+fake = Faker()
+
+
+class PopulatorTestCase(test.TestCase):
+
+    def testPopulation(self):
+        generator = fake
+        populator = Populator(generator)
+        populator.add_entity(Game, 10)
+        self.assertEqual(len(populator.execute()[Game]), 10)
+
+    def testGuesser(self):
+        generator = fake
+
+        def title_fake(arg):
+            title_fake.count += 1
+            name = generator.company()
+            return name
+
+        title_fake.count = 0
+
+        populator = Populator(generator)
+        populator.add_entity(Game, 10, {
+            'title': title_fake
+        })
+        self.assertEqual(len(populator.execute()[Game]), title_fake.count)
+
+    def testFormatter(self):
+        generator = fake
+
+        populator = Populator(generator)
+
+        populator.add_entity(Game, 5)
+        populator.add_entity(Player, 10, {
+            'score': lambda x: fake.random_int(0, 1000),
+            'nickname': lambda x: fake.email()
+        })
+        populator.add_entity(Action, 30)
+
+        insertedPks = populator.execute()
+
+        self.assertTrue(len(insertedPks[Game]) == 5)
+        self.assertTrue(len(insertedPks[Player]) == 10)
+
+        self.assertTrue(any([0 <= p.score <= 1000 and '@' in p.nickname for p in Player.objects.all()]))
 
 
 class ExampleTestCase(test.TestCase):
@@ -11,7 +59,7 @@ class ExampleTestCase(test.TestCase):
         populator.clear()
 
     def test_complex(self):
-        """Should be able to populate a complex database"""
+        """Should be able to populate a complex set of models"""
         populator = DjangoFaker.get_populator()
         populator.add_entity(Assessment, 5)
         populator.add_entity(Question, 5)
